requirements-test.txt
# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0
factory-boy==3.3.0
freezegun==1.2.2  # For datetime mocking